Apply resilience across

+++ BigBang1112.WorldRecordReportLib/Repos/IMapRepo.cs
@@ -7,7 +7,7 @@ public interface IMapRepo : IRepo<MapModel>
     Task<MapModel?> GetByUidAsync(string mapUid, CancellationToken cancellationToken = default);
     Task<Guid?> GetMapIdByMapUidAsync(string mapUid, CancellationToken cancellationToken = default);
     Task<List<MapModel>> GetByCampaignAsync(CampaignModel campaign, CancellationToken cancellationToken = default);
-    Task<MapModel?> GetByMxIdAsync(int trackId, TmxSite tmxSite, CancellationToken cancellationToken = default);
+    Task<MapModel?> GetByMxIdAsync(long trackId, TmxSite tmxSite, CancellationToken cancellationToken = default);
     Task<IEnumerable<MapModel>> GetByMultipleParamsAsync(string? mapName = null, string? env = null, string? title = null, string? authorLogin = null, string? authorNickname = null, int limit = 25, CancellationToken cancellationToken = default);
     Task<IEnumerable<MapModel>> GetByCampaignsThatAreOverAsync(Game game, CancellationToken cancellationToken = default);
     Task<IEnumerable<MapModel>> GetAllByCampaignLeaderboardUidAsync(string leaderboardUid, CancellationToken cancellationToken = default);
